pyproject.toml
[tool.poetry]
name = "predator-analytics-v13"
version = "13.0.0"
description = "Autonomous MAS Platform for Customs/Tax/OSINT Analytics with 30+ Agents, 58 LLM Models, Voice Interface, Self-Learning"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
celery = {extras = ["redis"], version = "^5.3.4"}
pandas = "^2.2.0"
pyarrow = "^15.0.0"
openpyxl = "^3.1.2"
pdfplumber = "^0.11.0"
pillow = "^10.2.0"
telethon = "^1.34.0"
playwright = "^1.41.0"
scrapy = "^2.11.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
opensearch-py = "^2.4.2"
qdrant-client = "^1.7.3"
openai = "^1.10.0"
anthropic = "^0.18.1"
cohere = "^4.47"
google-generativeai = "^0.3.2"
groq = "^0.4.2"
langchain = "^0.1.5"
langchain-community = "^0.0.38"
langchain-openai = "^0.0.2"
langgraph = "^0.0.26"
crewai = {extras = ["tools"], version = "^0.1.27"}
pyautogen = "^0.2.16"
mlflow = "^2.10.2"
transformers = "^4.37.2"
torch = "^2.2.0"
peft = "^0.8.2"
datasets = "^2.16.1"
spacy = "^3.7.2"
great-expectations = "^0.18.8"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
python-multipart = "^0.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.26.0"
aiofiles = "^23.2.1"
websockets = "^12.0"
aiokafka = "^0.10.0"
pydub = "^0.25.1"
pyttsx3 = "^2.90"
soundfile = "^0.12.1"
librosa = "^0.10.1"
neo4j = "^5.16.0"
networkx = "^3.2.1"
plotly = "^5.18.0"
kaleido = "0.2.1"
prophet = "^1.1.5"
lightgbm = "^4.3.0"
xgboost = "^2.0.3"
scikit-learn = "^1.4.0"
statsmodels = "^0.14.1"
python-dotenv = "^1.0.1"
tenacity = "^8.2.3"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
markdown = "^3.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
faker = "^22.6.0"
locust = "^2.20.1"
black = "^24.1.1"
ruff = "^0.2.1"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501", "N806", "N812"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "--cov=. --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
